feat(alerting): a paused rule says when it stopped and who stopped it
The detail header printed "nothing will be sent · rule is not evaluated"
beside a state word that already read "Paused", so the line spent itself
restating the status. It now carries the pause trail instead: "since 14m
by Ada", and stays empty for rules paused before the trail existed.

`alert_definitions` grows a pause trail the way `alert_silences` carries
its author: the stamp, the canonical principal, and the display at pause
time, written and cleared as one fact so a rename cannot rewrite who the
trail names. The read side keeps the "since" preposition in `format.ts`,
next to the phrasings it has to compose with.

Also lands the read-side work this branch had open: the instance value
loading moves out of `history.ts` into `values.ts`, and the delivery
lookup becomes one ordered index probe per rule.
