+++ packages/app/drizzle/0013_good_malice.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "alert_definitions" ADD COLUMN "paused_by" text;
