+++ packages/app/src/data/alerting/resource/schema.ts
@@ -1,4 +1,4 @@
-import * as z from "zod";
+import * as z from "zod/v4";
 import { dashboardProjectSchema } from "@/data/dashboards/schema";
 import { isReservedAnnotationKey } from "../resource-annotations";
